requirements.txt
bcrypt==3.1.3
cffi==1.10.0
click==6.7
Flask==0.12.1
Flask-Bcrypt==0.7.1
Flask-Login==0.4.0
Flask-WTF==0.14.2
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
peewee==2.10.1
pycparser==2.17
python-slugify==1.2.4
six==1.10.0
Unidecode==0.4.20
Werkzeug==0.12.1
WTForms==2.1
